fix: move equity universe metadata out of bounded package seed
