+++ research/equity_regions.py
